requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
drf-yasg>=1.21.10
celery>=5.3.6
redis>=5.0.1
Pillow>=10.2.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
requests 